requirements.txt
web3==6.20.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.5.2
torch==2.4.1
pycoingecko==3.2.0
python-dotenv==1.0.1